chore: Inline auth & jobs routes into the main router file
